+++ R/download.R
@@ -65,6 +65,10 @@ download_hydat <- function(dl_hydat_here = NULL) {
   ## Create the link to download HYDAT
   base_url <-
     "http://collaboration.cmc.ec.gc.ca/cmc/hydrometrics/www/"
+  
+  # Run network check
+  network_check(base_url)
+  
   x <- httr::GET(base_url)
   httr::stop_for_status(x)
   new_hydat <- substr(gsub("^.*\\Hydat_sqlite3_", "",
@@ -92,7 +96,8 @@ download_hydat <- function(dl_hydat_here = NULL) {
   #if(!dir.exists(dirname(tmp))) dir.create(dirname(tmp))
   
   ## Download the zip file
-  res <- httr::GET(url, httr::write_disk(tmp), httr::progress("down"))
+  res <- httr::GET(url, httr::write_disk(tmp), httr::progress("down"), 
+                   httr::user_agent("https://github.com/ropensci/tidyhydat"))
   on.exit(file.remove(tmp))
   httr::stop_for_status(res)