+++ R/hy_monthly_flows.R
@@ -12,7 +12,7 @@
 
 #' Extract monthly flows information from the HYDAT database
 #'
-#' Tidy data of monthly flows information from the DLY_FLOWS HYDAT table. \code{station_number} and
+#' Tidy data of monthly flows information from the monthly_flows HYDAT table. \code{station_number} and
 #' \code{prov_terr_state_loc} can both be supplied. If both are omitted all values from the \code{hy_stations} table are returned.
 #' That is a large vector for \code{hy_monthly_flows}.
 #'
@@ -25,10 +25,10 @@
 #' @format A tibble with 8 variables:
 #' \describe{
 #'   \item{STATION_NUMBER}{Unique 7 digit Water Survey of Canada station number}
-#'   \item{YEAR}{Year of record.}
-#'   \item{MONTH}{Numeric month value}
-#'   \item{FULL_MONTH}{Logical value is there is full record from MONTH}
-#'   \item{NO_DAYS}{Number of days in that month}
+#'   \item{Year}{Year of record.}
+#'   \item{Month}{Numeric month value}
+#'   \item{Full_Month}{Logical value is there is full record from Month}
+#'   \item{No_days}{Number of days in that month}
 #'   \item{Sum_stat}{Summary statistic being used.}
 #'   \item{Value}{Value of the measurement in m^3/s.}
 #'   \item{Date_occurred}{Observation date. Formatted as a Date class. MEAN is a annual summary
@@ -51,7 +51,12 @@
 
 hy_monthly_flows <- function(station_number = NULL,
                              hydat_path = NULL,
-                             prov_terr_state_loc = NULL, start_date ="ALL", end_date = "ALL") {
+                             prov_terr_state_loc = NULL, 
+                             start_date =NULL, 
+                             end_date = NULL) {
+  
+  ## Determine which dates should be queried
+  dates_null <- date_check(start_date, end_date)
 
   ## Read in database
   hydat_con <- hy_src(hydat_path)
@@ -59,37 +64,11 @@ hy_monthly_flows <- function(station_number = NULL,
     on.exit(hy_src_disconnect(hydat_con))
   }
 
-  if (start_date == "ALL" & end_date == "ALL") {
-    message("No start and end dates specified. All dates available will be returned.")
-  } else {
-    ## When we want date contraints we need to break apart the dates because SQL has no native date format
-    ## Start
-    start_year <- lubridate::year(start_date)
-    start_month <- lubridate::month(start_date)
-    start_day <- lubridate::day(start_date)
-
-    ## End
-    end_year <- lubridate::year(end_date)
-    end_month <- lubridate::month(end_date)
-    end_day <- lubridate::day(end_date)
-  }
-
-  ## Check date is in the right format
-  if (start_date != "ALL" | end_date != "ALL") {
-    if (is.na(as.Date(start_date, format = "%Y-%m-%d")) | is.na(as.Date(end_date, format = "%Y-%m-%d"))) {
-      stop("Invalid date format. Dates need to be in YYYY-MM-DD format")
-    }
-
-    if (start_date > end_date) {
-      stop("start_date is after end_date. Try swapping values.")
-    }
-  }
-
   ## Determine which stations we are querying
   stns <- station_choice(hydat_con, station_number, prov_terr_state_loc)
 
   ## Creating rlang symbols
-  sym_YEAR <- sym("YEAR")
+  sym_YEAR <- sym("Year")
   sym_STATION_NUMBER <- sym("STATION_NUMBER")
   sym_variable <- sym("variable")
   sym_temp <- sym("temp")
@@ -100,43 +79,43 @@ hy_monthly_flows <- function(station_number = NULL,
   monthly_flows <- dplyr::filter(monthly_flows, !!sym_STATION_NUMBER %in% stns)
 
   ## Do the initial subset to take advantage of dbplyr only issuing sql query when it has too
-  if (start_date != "ALL" | end_date != "ALL") {
-    monthly_flows <- dplyr::filter(monthly_flows, !!sym_YEAR >= start_year &
-      !!sym_YEAR <= end_year)
-
-    # monthly_flows <- dplyr::filter(monthly_flows, MONTH >= start_month &
-    #                             MONTH <= end_month)
-  }
+  
+  ## by year
+  if (!dates_null[["start_is_null"]]) monthly_flows <- dplyr::filter(monthly_flows, !!sym_YEAR >= lubridate::year(start_date))
+  if (!dates_null[["end_is_null"]]) monthly_flows <- dplyr::filter(monthly_flows, !!sym_YEAR <= lubridate::year(end_date))
 
   monthly_flows <- dplyr::select(monthly_flows, .data$STATION_NUMBER:.data$MAX)
   monthly_flows <- dplyr::collect(monthly_flows)
 
-  if (is.data.frame(monthly_flows) && nrow(monthly_flows) == 0) {
-    stop("This station is not present in HYDAT")
-  }
+  if (is.data.frame(monthly_flows) && nrow(monthly_flows) == 0) stop("This station is not present in HYDAT")
+
 
   ## Need to rename columns for gather
   colnames(monthly_flows) <- c(
-    "STATION_NUMBER", "YEAR", "MONTH", "FULL_MONTH", "NO_DAYS", "MEAN_Value",
+    "STATION_NUMBER", "Year", "Month", "Full_Month", "No_days", "MEAN_Value",
     "TOTAL_Value", "MIN_DAY", "MIN_Value", "MAX_DAY", "MAX_Value"
   )
 
 
 
-  monthly_flows <- tidyr::gather(monthly_flows, !!sym_variable, !!sym_temp, -(.data$STATION_NUMBER:.data$NO_DAYS))
+  monthly_flows <- tidyr::gather(monthly_flows, !!sym_variable, !!sym_temp, -(.data$STATION_NUMBER:.data$No_days))
   monthly_flows <- tidyr::separate(monthly_flows, !!sym_variable, into = c("Sum_stat", "temp2"), sep = "_")
 
   monthly_flows <- tidyr::spread(monthly_flows, !!sym_temp2, !!sym_temp)
 
   ## convert into R date for date of occurence.
-  monthly_flows <- dplyr::mutate(monthly_flows, Date_occurred = paste0(.data$YEAR, "-", .data$MONTH, "-", .data$DAY))
+  monthly_flows <- dplyr::mutate(monthly_flows, Date_occurred = paste0(.data$Year, "-", .data$Month, "-", .data$DAY))
   
   ## Check if DAY is NA and if so give it an NA value so the date parse correctly.
   monthly_flows <- dplyr::mutate(monthly_flows, Date_occurred = ifelse(is.na(.data$DAY), NA, .data$Date_occurred))
   monthly_flows <- dplyr::mutate(monthly_flows, Date_occurred = lubridate::ymd(.data$Date_occurred, quiet = TRUE))
+  
+  ## Then when a date column exist fine tune the subset
+  if (!dates_null[["start_is_null"]]) monthly_flows <- dplyr::filter(monthly_flows, .data$Date_occurred >= start_date)
+  if (!dates_null[["end_is_null"]]) monthly_flows <- dplyr::filter(monthly_flows, .data$Date_occurred <= end_date)
 
   monthly_flows <- dplyr::select(monthly_flows, -.data$DAY)
-  monthly_flows <- dplyr::mutate(monthly_flows, FULL_MONTH = .data$FULL_MONTH == 1)
+  monthly_flows <- dplyr::mutate(monthly_flows, Full_Month = .data$Full_Month == 1)
 
   ## What stations were missed?
   differ_msg(unique(stns), unique(monthly_flows$STATION_NUMBER))